+++ WORKSPACE
@@ -1,11 +1,16 @@
 # The nodejs rules
 RULES_NODEJS_VERSION = "2.0.0"
+
 RULES_NODEJS_SHA256 = "5bf77cc2d13ddf9124f4c1453dd96063774d755d4fc75d922471540d1c9a8ea8"
+
 # Rules for compiling sass
 RULES_SASS_VERSION = "1.26.3"
+
 RULES_SASS_SHA256 = "9dcfba04e4af896626f4760d866f895ea4291bc30bf7287887cefcf4707b6a62"
+
 # Bazel toolchain needed for remote execution
 BAZEL_TOOLCHAIN_VERSION = "3.2.0"
+
 BAZEL_TOOLCHAIN_SHA256 = "db48eed61552e25d36fe051a65d2a329cc0fb08442627e8f13960c5ab087a44e"
 
 workspace(
@@ -15,7 +20,6 @@ workspace(
 
 # These rules are built-into Bazel but we need to load them first to download more rules
 load("@bazel_tools//tools/build_defs/repo:http.bzl", "http_archive")
-
 
 http_archive(
     name = "build_bazel_rules_nodejs",
@@ -37,8 +41,10 @@ http_archive(
 # Npm Install and Typescript Sass support #
 ###########################################
 load("@build_bazel_rules_nodejs//:index.bzl", "check_bazel_version", "npm_install")
+
 # The minimum bazel version to use with this repo is v3.1.0.
 check_bazel_version("3.1.0")
+
 # Install all dependencies with npm
 npm_install(
     name = "npm",
@@ -58,6 +64,7 @@ install_bazel_dependencies(suppress_warning = True)
 load("@io_bazel_rules_sass//sass:sass_repositories.bzl", "sass_repositories")
 
 sass_repositories()
+
 # Fetch required transitive dependencies. This is an optional step because you
 # can always fetch the required NodeJS transitive dependency on your own.
 load("@io_bazel_rules_sass//:package.bzl", "rules_sass_dependencies")
