--- apps/design-system/_data/navitems.js
@@ -1,16 +1,0 @@
-module.exports = function () {
-  return [
-    {
-      url: '/design/',
-      label: 'Design',
-    },
-    {
-      url: '/develop/',
-      label: 'Develop',
-    },
-    {
-      url: '/components/',
-      label: 'Components',
-    },
-  ];
-};
