+++ apps/design-system/src/component.njk
