+++ apps/design-system/src/data/BUILD.bazel
@@ -1,0 +1,20 @@
+load("@npm//@bazel/typescript:index.bzl", "ts_library")
+
+package(default_visibility = ["//apps/design-system:__subpackages__"])
+
+ts_library(
+    name = "data",
+    srcs = glob(
+        include = ["**/*.ts"],
+    ),
+    tsconfig = "//apps/design-system:tsconfig_data",
+    devmode_module = "commonjs",
+    deps = [
+        "@npm//@types/node",
+        "@npm//@types/glob",
+        "@npm//axios",
+        "@npm//glob",
+        "@npm//gray-matter",
+        "@npm//tslib",
+    ],
+)
