+++ apps/design-system/src/non-component-pages.njk
@@ -9,4 +9,3 @@ pagination:
   addAllPagesToCollections: true
 permalink: "{{ post.slug }}/index.html"
 ---
-
