+++ apps/design-system/src/styles/BUILD.bazel
@@ -1,0 +1,16 @@
+load("@io_bazel_rules_sass//:defs.bzl", "sass_binary", "sass_library")
+
+package(default_visibility = ["//apps/design-system:__subpackages__"])
+
+sass_library(
+    name = "scss",
+    srcs = glob(
+        ["**/_*.scss"],
+    ),
+)
+
+sass_binary(
+    name = "styles",
+    src = "main.scss",
+    deps = [":scss"],
+)
