+++ apps/design-system/src/styles/_base.scss
