+++ apps/design-system/src/styles/_sidebar.scss
