+++ apps/design-system/src/styles/_top-navigation.scss
