+++ apps/design-system/src/styles/_variables.scss
