+++ apps/design-system/src/templates/layouts/component.njk
