+++ apps/design-system/src/templates/layouts/layout-variants/with-sidebar.njk
