+++ apps/design-system/src/templates/layouts/layout-variants/without-sidebar.njk
