+++ apps/design-system/src/templates/layouts/page.njk
@@ -7,3 +7,4 @@ layout: layouts/base.njk
 {% block main %}
   {{ post.content | markdownify | safe }}
 {% endblock %}
+
