+++ apps/design-system/tsconfig.app.json
@@ -1,0 +1,10 @@
+{
+  "extends": "./tsconfig.json",
+  "compilerOptions": {
+    "outDir": "../../dist/out-tsc",
+    "types": []
+  },
+  "files": ["scripts/main.ts"],
+  "include": ["**/*.d.ts"],
+  "exclude": ["**/*.spec.ts", "**/*.e2e.ts", "**/*.po.ts"]
+}
