+++ apps/design-system/tsconfig.json
@@ -1,8 +1,13 @@
 {
-  "extends": "../../tsconfig.json",
-  "compilerOptions": {
-    "outDir": "../../out-tsc",
-    "module": "commonjs"
-  },
-  "include": ["**/*.ts"]
+  "extends": "../../tsconfig.base.json",
+  "include": [],
+  "files": [],
+  "references": [
+    {
+      "path": "./tsconfig.app.json"
+    },
+    {
+      "path": "./tsconfig.data.json"
+    }
+  ]
 }
