+++ package.json
@@ -150,6 +150,7 @@
     "css-loader": "^3.5.3",
     "d3-array": "^2.4.0",
     "dotenv": "^8.2.0",
+    "esbuild": "^0.6.21",
     "express": "^4.15.2",
     "glob": "^7.1.6",
     "gonzales-pe": "^4.3.0",
