+++ tsconfig.base.json
@@ -6,6 +6,7 @@
     "declaration": false,
     "emitDecoratorMetadata": true,
     "experimentalDecorators": true,
+    "esModuleInterop": true,
     "importHelpers": true,
     "lib": ["es2018", "dom"],
     "module": "esnext",
